:wrench: Consolidate open-note with preview epic
Both related and needs handling the preview
