+++ lib/epics/index.js
@@ -6,7 +6,6 @@ import configChanges from './config-changes'
 import makeHiddenColumns from './hidden-columns'
 import makeFileReads from './file-reads'
 import makeFileWrites from './file-writes'
-import openNote from './open-note'
 import pathWatcher from './path-watcher'
 import previewNote from './preview-note'
 
@@ -22,7 +21,6 @@ export default function makeEpicMiddleware (columns: Columns, fileReaders: FileR
       fileReads,
       fileWrites,
       hiddenColumns,
-      openNote,
       pathWatcher,
       previewNote
     )
