+++ lib/epics/preview-note.js
@@ -6,6 +6,7 @@ import PreviewElement from '../preview-element'
 import * as A from '../action-creators'
 
 const PREVIEW_SCHEMA_PREFIX = 'tv://'
+const HAS_FILE_EXT_REGEX = /\.\w{1,5}$/
 
 export default function previewNoteEpic (action$: rxjs$Observable<Action>, store: Store<State, Action>) {
   const previewElement = new PreviewElement()
@@ -73,6 +74,33 @@ export default function previewNoteEpic (action$: rxjs$Observable<Action>, store
           return false
         }),
 
+      // Open on specific open-note action
+      action$
+        .filter(action => {
+          if (action.type === A.OPEN_NOTE) {
+            const state: State = store.getState()
+            let filename
+
+            if (state.selectedNote) {
+              filename = state.selectedNote.filename
+            } else {
+              filename = state.queryOriginal.trim() || 'untitled'
+
+              if (!HAS_FILE_EXT_REGEX.test(filename)) {
+                const ext = atom.config.get('textual-velocity.defaultExt')
+                  .replace(/^\./, '') // avoid double dots next to extension, i.e. untitled..txt => untitled.txt
+                filename = `${filename}.${ext}`
+              }
+            }
+
+            atom.workspace
+              .open(Path.join(state.dir, filename))
+              .then(closePreview)
+          }
+
+          return false
+        }),
+
       // Replace preview with default editor when clicked
       Observable
         .fromEvent(previewElement, 'click')
@@ -82,16 +110,16 @@ export default function previewNoteEpic (action$: rxjs$Observable<Action>, store
             .then(closePreview)
           return false
         })
-    )
+    ) // .merge
     .takeUntil(
       action$
         .filter(action => {
-          if (action.type === A.DISPOSE) {
+          const isDisposeAction = action.type === A.DISPOSE
+          if (isDisposeAction) {
             closePreview()
             openerDisposable.dispose()
             previewElement.dispose()
-            return true
           }
-          return false
+          return isDisposeAction
         }))
 }
